+++ lib/strong_migrations.rb
@@ -29,7 +29,7 @@ module StrongMigrations
       :target_postgresql_version, :target_mysql_version, :target_mariadb_version,
       :enabled_checks, :lock_timeout, :statement_timeout, :check_down, :target_version,
       :safe_by_default, :target_sql_mode, :lock_timeout_retries, :lock_timeout_retry_delay,
-      :alphabetize_schema
+      :alphabetize_schema, :ignored_databases
     attr_writer :lock_timeout_limit
   end
   self.auto_analyze = false
@@ -40,6 +40,7 @@ module StrongMigrations
   self.safe_by_default = false
   self.check_down = false
   self.alphabetize_schema = false
+  self.ignored_databases = []
 
   # private
   def self.developer_env?