+++ lib/strong_migrations/migrator.rb
@@ -6,8 +6,10 @@ module StrongMigrations
       # handle MigrationProxy class
       migration = migration.send(:migration) if migration.respond_to?(:migration, true)
 
-      # retry migration since the entire transaction needs to be rerun
       checker = migration.send(:strong_migrations_checker)
+      return super unless checker.send(:enabled?)
+
+      # retry migration since the entire transaction needs to be rerun
       checker.retry_lock_timeouts(check_committed: true) do
         # failed transaction reverts timeout, so need to re-apply
         checker.timeouts_set = false